Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -7,7 +7,7 @@ float main(){
     scanf("%f",&b);
     printf("");
     scanf("%f",&c);
-    int avg = (a+b+c)/3.0;
+    float avg = (a+b+c)/3.0;
     printf("Average: %.2f\n", avg);
     return 0;
 }
